Consolidate teaser tests into inline-media-link-editing and fix readonly behavior
- Move all teaser starter UI tests from teaser-starter-ui.spec.ts to
  inline-media-link-editing.spec.ts (they test same functionality)
- Delete teaser-starter-ui.spec.ts
- Fix test frontend renderer to add data-block-readonly when teaser
  overwrite is false (prevents editing until user clicks Customize)
- Update test to check behavior (can't edit) not implementation detail
