--- tests-playwright/integration/teaser-starter-ui.spec.ts
@@ -1,292 +1,0 @@
-/**
- * Tests for teaser block starter UI.
- * When a teaser block has an empty required 'href' field, a starter UI overlay
- * should appear with an AddLinkForm to help the user select a target page.
- */
-import { test, expect } from '../fixtures';
-import { AdminUIHelper } from '../helpers/AdminUIHelper';
-
-test.describe('Teaser Starter UI', () => {
-  test('shows starter UI overlay for empty teaser href field', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Find and click the empty teaser block
-    const emptyTeaser = iframe.locator('[data-block-uid="block-6-empty-teaser"]');
-    await expect(emptyTeaser).toBeVisible({ timeout: 10000 });
-    await emptyTeaser.click();
-
-    // Wait for block to be selected (outline appears)
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // The starter UI overlay should appear with link icon and AddLinkForm
-    const starterOverlay = page.locator('.starter-ui-overlay');
-    await expect(starterOverlay).toBeVisible({ timeout: 5000 });
-
-    // Should have the link form input
-    const linkInput = starterOverlay.locator('input[name="link"]');
-    await expect(linkInput).toBeVisible();
-  });
-
-  test('does not show starter UI for teaser with href filled', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Find and click the filled teaser block
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    await expect(filledTeaser).toBeVisible({ timeout: 10000 });
-    await filledTeaser.click();
-
-    // Wait for block to be selected
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // Starter UI should NOT appear (teaser has href filled)
-    const starterOverlay = page.locator('.starter-ui-overlay');
-    await expect(starterOverlay).not.toBeVisible();
-  });
-
-  test('teaser shows target page title and description from href', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Find the filled teaser block
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    await expect(filledTeaser).toBeVisible({ timeout: 10000 });
-
-    // Teaser should show title from href[0].title (not a custom block.title)
-    // Use h3, h5 to support both mock (h3) and Nuxt (h5) frontends
-    const title = filledTeaser.locator('h3, h5');
-    await expect(title).toHaveText('Target Page');
-
-    // Teaser should show description from href[0].description
-    const description = filledTeaser.locator('p').first();
-    await expect(description).toHaveText('Target page description');
-  });
-
-  test('teaser title becomes editable after clicking customize checkbox and reverts when unchecked', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Click the filled teaser to select it
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    await expect(filledTeaser).toBeVisible({ timeout: 10000 });
-    await filledTeaser.click();
-
-    // Wait for block to be selected
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // Initially overwrite is false, so block should have data-block-readonly
-    // hydra.js ignores editable fields inside readonly blocks
-    await expect(filledTeaser).toHaveAttribute('data-block-readonly');
-
-    // Title element should have data-editable-field but NOT be contenteditable (readonly)
-    const titleLocator = filledTeaser.locator('[data-editable-field="title"]');
-    await expect(titleLocator).toHaveCount(1);
-    await expect(titleLocator).not.toHaveAttribute('contenteditable', 'true');
-
-    // Verify title shows target's title initially (h3 in mock, h5 in Nuxt)
-    const titleElement = filledTeaser.locator('h3, h5');
-    await expect(titleElement).toHaveText('Target Page');
-
-    // Click the "Customize teaser content" checkbox in sidebar
-    const customizeCheckbox = page.locator('text=Customize teaser content');
-    await expect(customizeCheckbox).toBeVisible({ timeout: 5000 });
-    await customizeCheckbox.click();
-
-    // After enabling customize, block should NOT have data-block-readonly
-    // and title should be contenteditable
-    // Wait for async render to complete (test frontend uses async expandAllListings)
-    await expect(filledTeaser).not.toHaveAttribute('data-block-readonly', { timeout: 5000 });
-    await expect(titleLocator).toHaveAttribute('contenteditable', 'true', { timeout: 5000 });
-
-    // Edit the title to something custom
-    await titleLocator.click();
-    await titleLocator.fill('Custom Title');
-    await expect(titleLocator).toHaveText('Custom Title');
-
-    // Click the checkbox again to uncustomize
-    await customizeCheckbox.click();
-
-    // After disabling customize:
-    // 1. Title should revert to target's title
-    await expect(titleElement).toHaveText('Target Page', { timeout: 5000 });
-
-    // 2. Block should have data-block-readonly again
-    await expect(filledTeaser).toHaveAttribute('data-block-readonly');
-
-    // 3. Click directly on the title - it should NOT become contenteditable
-    // This is the key bug: clicking on title after unchecking should not make it editable
-    await titleElement.click();
-    await expect(titleElement).not.toHaveAttribute('contenteditable', 'true');
-  });
-
-  test('can select target in starter UI and fill teaser href', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Click the empty teaser (scroll into center of view so overlay has room)
-    const emptyTeaser = iframe.locator('[data-block-uid="block-6-empty-teaser"]');
-    await expect(emptyTeaser).toBeVisible({ timeout: 10000 });
-    await emptyTeaser.scrollIntoViewIfNeeded();
-    await page.waitForTimeout(200); // Wait for scroll to settle
-    await emptyTeaser.click();
-
-    // Wait for starter UI
-    const starterOverlay = page.locator('.starter-ui-overlay');
-    await expect(starterOverlay).toBeVisible({ timeout: 5000 });
-
-    // Open object browser from starter UI (use force since it may be at edge of viewport)
-    const browseButton = starterOverlay.locator('button[aria-label="Open object browser"]');
-    await expect(browseButton).toBeVisible({ timeout: 5000 });
-    await browseButton.click({ force: true });
-    const objectBrowser = await helper.waitForObjectBrowser();
-
-    // Select "Another Page" from the object browser (helper closes it if needed)
-    await helper.objectBrowserSelectItem(objectBrowser, /Another Page/);
-
-    // Submit the AddLinkForm if still open
-    await helper.submitAddLinkFormIfOpen(starterOverlay);
-
-    // After submitting, the starter UI should disappear (href is now filled)
-    await expect(starterOverlay).not.toBeVisible({ timeout: 5000 });
-
-    // Verify the teaser now has the "Read more" link with correct href
-    // Note: Title/description aren't shown because AddLinkForm only sets @id,
-    // not the full metadata. This would require the object_browser widget to
-    // resolve the target's metadata, which is a future enhancement.
-    const readMoreLink = emptyTeaser.locator('a').first();
-    await expect(readMoreLink).toBeVisible({ timeout: 5000 });
-    await expect(readMoreLink).toHaveAttribute('href', /another-page/);
-  });
-
-  test('clicking Read more link does not navigate and shows link icon', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Find and click the filled teaser to select it first
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    await expect(filledTeaser).toBeVisible({ timeout: 10000 });
-    await filledTeaser.click();
-
-    // Wait for block to be selected
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // Enable customize mode so linkable fields become editable
-    const customizeCheckbox = page.locator('text=Customize teaser content');
-    await expect(customizeCheckbox).toBeVisible({ timeout: 5000 });
-    await customizeCheckbox.click();
-
-    // Wait for overwrite to be enabled (data-block-readonly removed)
-    await expect(filledTeaser).not.toHaveAttribute('data-block-readonly', { timeout: 5000 });
-
-    // Now click the "Read more" link
-    const readMoreLink = filledTeaser.locator('a[data-linkable-field="href"]').last();
-    await expect(readMoreLink).toBeVisible();
-
-    // Get current URL before clicking
-    const urlBefore = page.url();
-
-    // Click the link
-    await readMoreLink.click();
-
-    // URL should NOT change (navigation prevented)
-    expect(page.url()).toBe(urlBefore);
-
-    // Link icon should appear in toolbar for editing href
-    const linkIcon = page.locator('button[title="Edit link (href)"]');
-    await expect(linkIcon).toBeVisible({ timeout: 5000 });
-  });
-
-  test('clicking teaser title shows link icon for editing href', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // Find and click the filled teaser to select it first
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    await expect(filledTeaser).toBeVisible({ timeout: 10000 });
-    await filledTeaser.click();
-
-    // Wait for block to be selected
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // Enable customize mode so linkable fields become editable
-    const customizeCheckbox = page.locator('text=Customize teaser content');
-    await expect(customizeCheckbox).toBeVisible({ timeout: 5000 });
-    await customizeCheckbox.click();
-
-    // Wait for overwrite to be enabled (data-block-readonly removed)
-    await expect(filledTeaser).not.toHaveAttribute('data-block-readonly', { timeout: 5000 });
-
-    // Now click the title link
-    const titleLink = filledTeaser.locator('a[data-linkable-field="href"]').first();
-    await expect(titleLink).toBeVisible();
-
-    // Get current URL before clicking
-    const urlBefore = page.url();
-
-    // Click the title
-    await titleLink.click();
-
-    // URL should NOT change (navigation prevented)
-    expect(page.url()).toBe(urlBefore);
-
-    // Link icon should appear in toolbar for editing href
-    const linkIcon = page.locator('button[title="Edit link (href)"]');
-    await expect(linkIcon).toBeVisible({ timeout: 5000 });
-  });
-
-  test('clicking Read more link on unselected block does not navigate', async ({ page }) => {
-    const helper = new AdminUIHelper(page);
-    await helper.login();
-    await helper.navigateToEdit('/test-page');
-
-    const iframe = helper.getIframe();
-
-    // First, click a different block to ensure teaser is NOT selected
-    const slateBlock = iframe.locator('[data-block-uid="block-1-uuid"]');
-    await expect(slateBlock).toBeVisible({ timeout: 10000 });
-    await slateBlock.click();
-
-    // Wait for that block to be selected
-    const outline = page.locator('.volto-hydra-block-outline');
-    await expect(outline).toBeVisible({ timeout: 5000 });
-
-    // Get current URL before clicking the teaser link
-    const urlBefore = page.url();
-
-    // Now click the "Read more" link in the teaser (which is NOT selected)
-    const filledTeaser = iframe.locator('[data-block-uid="block-7-filled-teaser"]');
-    const readMoreLink = filledTeaser.locator('a[data-linkable-field="href"]').last();
-    await expect(readMoreLink).toBeVisible();
-    await readMoreLink.click();
-
-    // URL should NOT change (navigation should be prevented even for unselected block)
-    expect(page.url()).toBe(urlBefore);
-  });
-});
